ProblemTextToJsonConverterをobsoleteに．
